generate AWS CFT from URL
This takes the release CFT and modifies it so that there is an image ID
parameter.  A default is set based on the tester configuration which
causes the configured image ID to be used.  In the future, the URL to
the release CFT should be configurable and the image ID should be set
using the parameter rather than hardcoded into the modified CFT.

Delete the cloudyClusterCloudFormationTemplate.json since it is now
completely generated.